+++ OUDS/Core/Tokens/SemanticTokens/Tests/Composites/MultipleSizingSemanticTokenTests.swift
@@ -14,19 +14,29 @@
 import XCTest
 import OUDSTokensSemantic
 
+<<<<<<< HEAD
 /// To ensure the `MultipleSizingTokens` is tested as a wrapper of semantic tokens for compact and regular size classes.
 final class MultipleSizingTokensTests: XCTestCase {
+=======
+/// To ensure the `MultipleSizingSemanticToken` is tested as a wrapper of semantic tokens for compact and regular size classes.
+final class MultipleSizingSemanticTokenTests: XCTestCase {
+>>>>>>> 3e552ea8 (refactor: use of composite colors for elevations)
 
     /// Tests if compact and regular values are preserved when defined
     func testInit() {
         let compact: SizingSemanticToken = 123
         let regular: SizingSemanticToken = 456
+<<<<<<< HEAD
         let token = MultipleSizingTokens(compact: compact, regular: regular)
+=======
+        let token = MultipleSizingSemanticToken(compact: compact, regular: regular)
+>>>>>>> 3e552ea8 (refactor: use of composite colors for elevations)
 
         XCTAssertTrue(token.compact == compact)
         XCTAssertTrue(token.regular == regular)
     }
 
+<<<<<<< HEAD
     /// Tests comparisons between two `MultipleSizingTokens` to ensure tokens are considered as equal
     /// if an only if they have the same compact and regular values and have the same types.
     func testIsEqual() {
@@ -36,6 +46,17 @@ final class MultipleSizingTokensTests: XCTestCase {
         let fourth = MultipleSizingTokens(compact: 56, regular: 34)
         let fifth = MultipleSizingTokens(compact: 12, regular: 34)
         let sixth = MultipleColorTokens("#000000")
+=======
+    /// Tests comparisons between two `MultipleSizingSemanticToken` to ensure tokens are considered as equal
+    /// if an only if they have the same compact and regular values and have the same types.
+    func testIsEqual() {
+        let first = MultipleSizingSemanticToken(compact: 12, regular: 34)
+        let second = MultipleSizingSemanticToken(compact: 56, regular: 78)
+        let third = MultipleSizingSemanticToken(compact: 12, regular: 78)
+        let fourth = MultipleSizingSemanticToken(compact: 56, regular: 34)
+        let fifth = MultipleSizingSemanticToken(compact: 12, regular: 34)
+        let sixth = MultipleColorSemanticToken("#000000")
+>>>>>>> 3e552ea8 (refactor: use of composite colors for elevations)
 
         XCTAssertTrue(first.isEqual(first))
         XCTAssertFalse(first.isEqual(second))